modules/connection_api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/production.in
#
aniso8601==9.0.1
    # via flask-restx
attrs==21.2.0
    # via jsonschema
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.6
    # via requests
click==8.0.2
    # via flask
colorama==0.4.4
    # via click
flask==2.0.2
    # via
    #   -r requirements\common.in
    #   flask-cors
    #   flask-restx
flask-accepts==0.18.4
    # via -r requirements\common.in
flask-cors==3.0.10
    # via -r requirements\common.in
flask-restx==0.5.1
    # via
    #   -r requirements\common.in
    #   flask-accepts
grpcio==1.41.0
    # via
    #   -r requirements\common.in
    #   grpcio-tools
grpcio-tools==1.41.0
    # via -r requirements\common.in
idna==3.2
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==3.0.2
    # via flask
jsonschema==4.0.1
    # via flask-restx
markupsafe==2.0.1
    # via jinja2
marshmallow==3.13.0
    # via
    #   -r requirements\common.in
    #   flask-accepts
protobuf==3.18.1
    # via grpcio-tools
pyrsistent==0.18.0
    # via jsonschema
python-dotenv==0.19.0
    # via -r requirements\common.in
pytz==2021.3
    # via flask-restx
requests==2.26.0
    # via -r requirements\common.in
six==1.16.0
    # via
    #   flask-cors
    #   flask-restx
    #   grpcio
urllib3==1.26.7
    # via requests
waitress==2.0.0
    # via -r requirements/production.in
werkzeug==2.0.2
    # via
    #   flask
    #   flask-accepts
    #   flask-restx

# The following packages are considered to be unsafe in a requirements file:
# setuptools
